Move schema files to src/crossPackage/, restructure output dirs, add CJS test
