+++ examples/utils/import-testing/src/crossPackage/schemaDefinitions.ts
@@ -4,14 +4,14 @@
  */
 
 /*
-Schema definitions that import the factory from a separate file (crossPackageSchemaUtils.ts).
+Schema definitions that import the factory from a separate file (schemaUtils.ts).
 This separation is critical: it forces the .d.ts to resolve tree/alpha types through a deeper
 chain, which triggers TypeScript's type resolution bug when tree's export tiers share a JS module.
 */
 
 /* eslint-disable jsdoc/require-jsdoc */
 
-import { sf } from "./crossPackageSchemaUtils.js";
+import { sf } from "./schemaUtils.js";
 
 export class Position extends sf.objectAlpha("Position", {
 	x: sf.number,
