+++ examples/utils/import-testing/src/crossPackage/schemaUtils.ts
