+++ examples/utils/import-testing/tsconfig.crossPackage.cjs.json
@@ -1,0 +1,7 @@
+{
+	// This config must be used in a "type": "commonjs" environment. (Use fluid-tsc commonjs.)
+	"extends": "./tsconfig.crossPackage.json",
+	"compilerOptions": {
+		"outDir": "./dist/crossPackage/node16",
+	},
+}
