+++ examples/utils/import-testing/tsconfig.crossPackageConsumer.cjs.json
@@ -1,0 +1,8 @@
+{
+	// This config must be used in a "type": "commonjs" environment. (Use fluid-tsc commonjs.)
+	"extends": "./tsconfig.crossPackageConsumer.json",
+	"include": ["src/test/crossPackageCjsImporter.ts"],
+	"compilerOptions": {
+		"outDir": "./dist/test/crossPackageConsumer",
+	},
+}
